The server now keeps track of database pool exhaustion events and will report daily how many occured, if some exhaustion was noticed. (Issue #5944, PR #6125)
# Description

Add the `ExhaustedPoolWatcher` class to count database pool  exhaustion events
- counter is increased when db exhaustion is detected (a recurring task is scheduled every 200ms in the `DatabaseService` startup)
- counter is logged as a warning every 24h and reset if any events were recorded (a recurring task is scheduled every 24h in the `DatabaseService` startup)

closes #5944

# Self Check:

Strike through any lines that are not applicable (`~~line~~`) then check the box

- [x] Attached issue to pull request
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors (check using make mypy or make mypy-diff)
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
~~- [ ] Correct, in line with design~~
~~- [ ] End user documentation is included or an issue is created for end-user documentation (add ref to issue here: )~~
~~- [ ] If this PR fixes a race condition in the test suite, also push the fix to the relevant stable branche(s) (see [test-fixes](https://internal.inmanta.com/development/core/tasks/build-master.html#test-fixes) for more info)~~

+++ src/inmanta/server/protocol.py
@@ -303,7 +303,7 @@ class ServerSlice(inmanta.protocol.endpoints.CallTarget, TaskHandler):
 
     # utility methods for extensions developers
     def schedule(
-        self, call: TaskMethod, interval: int = 60, initial_delay: Optional[float] = None, cancel_on_stop: bool = True
+        self, call: TaskMethod, interval: float = 60, initial_delay: Optional[float] = None, cancel_on_stop: bool = True
     ) -> None:
         """
         Schedule a task repeatedly with a given interval. Tasks with the same call and the same schedule are considered the
@@ -312,7 +312,7 @@ class ServerSlice(inmanta.protocol.endpoints.CallTarget, TaskHandler):
         :param interval: The interval between executions of the task.
         :param initial_delay: The delay to execute the task for the first time. If not set, interval is used.
         """
-        self._sched.add_action(call, IntervalSchedule(float(interval), initial_delay), cancel_on_stop)
+        self._sched.add_action(call, IntervalSchedule(interval, initial_delay), cancel_on_stop)
 
     def schedule_cron(self, call: TaskMethod, cron: str, cancel_on_stop: bool = True) -> None:
         """
